fix: :bug: correct threshold type hint

+++ feo/client/asset.py
@@ -14,7 +14,11 @@ class Asset(schemas.NodeBase):
 
     @classmethod
     def search(
-        cls, alias: str, threshold: int = 0.5, node_type: str = None, sector: str = None
+        cls,
+        alias: str,
+        threshold: float = 0.5,
+        node_type: str = None,
+        sector: str = None,
     ) -> List["schemas.Node"]:
         """
         Search for nodes using an alias.
